Style improvements.
Don't use a pointer for RepoUpdate struct since it's not neccessary.

Simplify docs to avoid stuttering.

Remove unneeded blank lines.

Move error handling closer to where error is being returned.

Remove trailing whitespace in protobuf file.

Remove unneccessary type conversion.

+++ cli/coverage_cmd.go
@@ -198,7 +198,7 @@ func ensureRepoExists(cl *sourcegraph.Client, ctx context.Context, repo string) 
 		// Automatically create a local mirror.
 		log15.Info("Creating a local mirror of remote repo", "cloneURL", remoteRepo.HTTPCloneURL)
 		_, err := cl.Repos.Create(ctx, &sourcegraph.ReposCreateOp{
-			Op: &sourcegraph.ReposCreateOp_FromGitHubID{FromGitHubID: int32(remoteRepo.GitHubID)},
+			Op: &sourcegraph.ReposCreateOp_FromGitHubID{FromGitHubID: remoteRepo.GitHubID},
 		})
 		if err != nil {
 			return err
